Unify transaction quick drawers
